input: goodix: Add option to poll instead of relying on IRQ line
The interrupt line from the touch controller is not necessarily
connected to the SoC, so add the option to poll for touch info.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

input: goodix: Include I2C details in names for the devices

libinput uses the input device name alone. If you have two
identical input devices, then there is no way to differentiate
between them, and in the case of touchscreens that means no
way to associate them with the appropriate display device.

Add the I2C bus and address to the start of the input device
name so that the name is always unique within the system.

Change-Id: I1b2ab2545fb9ffd1db7182e1647110c772acfd36
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>