drm: panel: ili9881: Add option to reconfigure setup commands
The driver is typically asking for LP commands, but then tries
to send set_display_[on|off] from enable/disable when the host
will be in HS mode.
It also sends shutdown commands just before it asserts reset and
disables the regulator, which is rather redundant.

Add an option to configure these two choices from the panel_desc.

Change-Id: I2d117f6337ecdcffb7673fbda5bd9f2f7c9cc1fd
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>