drm/panel: panel-simple: add panel_reinit property detection
reinit means re-initialization, due to the timing of some panel does not
in firmware, it will cause the panel need re-initialization from suspend
mode to wakeup. In this case, keep the power on after first initialized
is a better way.

So, separated two cases using a property named "no-panel-reinit" for
this solution.