arm64: configs: rockchip: Migrate configs for Bifrost and Valhall GPUs to SoC-level config files
This is in coordination with the modification of commit
"MALI: rockchip: Add separate src directory for Valhall driver from DDK g28p0-00eac0".
The configs related to Bifrost and Valhall GPUs are designed to be configured
in the SoC-level config files.
Therefore, for some SoCs, the kernel compilation command line needs to be adjusted.

Change-Id: Ida29d2f6dba9b9ea7f39e8b5f7995214b346d1dc
Signed-off-by: Zhen Chen <chenzhen@rock-chips.com>